pyproject.toml


[tool.hatch.build.targets.wheel]
packages = ["src/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hotel-reservation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.1.0",
    "imbalanced-learn>=0.13.0",
    "ipykernel>=6.29.5",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.1",
    "mlflow>=2.22.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "xgboost>=3.0.0",
]

[tool.setuptools]
packages = ["src"]
